Started Setting Base Mechanics
I have a 3D background working and made 2D spite assets for the plant placeholders, which contain a pot and a plant.

Added a GameManager Script. I'm currently thinking of having the GameManager randomly assign the pot and plant from a list/array. The GameManager would then also keep track of growth and when the plant can be used, as well as the customer's requests.

Tried making a Plant ScriptableObject, but it's not working; still trying to figure out why.