Added More Vegetable Types
Added the rough sketches of other vegetable types. Thinking of getting rid of the growing mechanic, so only add a stage 4 for some.

Really burnt out on the idea of this game and am stumped on where to take it from here. Growing is waiting, which is not good. The customer request isn't even a good way to do scoring. I could lean into the identification idea, but I don't know how to score that besides right/wrong.