Randomization and New Direction
Decided to get rid of customer requests and have to score be based on labeling the plants correctly.

Randomized the plants and the growth stages.

Need to complete the art and decide how I want to go about labeling the plants. Was thinking of doing a drag-and-drop, but might go with something else.