Added Info and Functionality to ID Book
Added plant descriptions and images to the ID book and made it, so you open it. Added a bit of code to make sure the ID book menu closes if the player tries to pause.

Decided not to make a strawberry plant.

Still needs a way to label plants and check if right/wrong.