Basic Mechanics and Functionality Seem Functional
Added a How to Play section with graphics. Fixed a typo in carrots description. Fixed a bug with the labels (the names of the gameobjects were wrong).

Going to go through my code to clean up and comment; playtest to find any potential bugs; update the ReadMe; Prettify the title menu??; etc.