Updated HTP, Clean Up, Bug Fixing
I updated the How to Play to include the newer mechanics. I cleaned up most of the code, but I might have missed some things.

I found a bug with the labels overlapping with the PlantSlots, so I made it so that they don't overlap. I also realized the DrapDown script was still on the PlantSlots causing the labels to stick to them too, and so I removed it. I also realized I forgot to reset the label names when the labels are reset, so I added that to the ResetLabels().